+++ testing/test-utils/src/main/resources/database-scripts/postgres/db-global-cleanup.sql
@@ -1,0 +1,3 @@
+DROP SCHEMA IF EXISTS "${schema}" CASCADE;
+DROP OWNED by "${schema}";
+DROP USER "${schema}";