+++ testing/test-utils/src/main/resources/database-scripts/postgres/db-setup.sql
@@ -1,0 +1,5 @@
+DROP SCHEMA IF EXISTS "${schema}" CASCADE;
+CREATE SCHEMA "${schema}";
+GRANT ALL ON SCHEMA "${schema}" TO "${schema}";
+ALTER DEFAULT privileges IN SCHEMA "${schema}" GRANT ALL ON tables TO "${schema}";
+ALTER DEFAULT privileges IN SCHEMA "${schema}" GRANT ALL ON sequences TO "${schema}";